+++ data-raw/irish_county_data.R
@@ -5,7 +5,6 @@ geo_irish_county <- sf::st_read("data-raw/geojson/counties_simple.geojson")
 irish_county_data <- readr::read_rds(
   'data-raw/rds/latest_irish_county_data.rds'
 ) %>%
-  dplyr::filter(Date == max(Date)) %>%
   dplyr::left_join(
     geo_irish_county,
     by = c("CountyName" = "NAME_TAG")